// frontend/src/api/generated/models/projectResource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */
import type { ProjectPriority } from './projectPriority.ts'
import type { ProjectStatus } from './projectStatus.ts'

export interface ProjectResource {
  id: number
  clientName: string
  projectName: string
  /** @nullable */
  description: string | null
  status: string | ProjectStatus
  priority: string | ProjectPriority
  /** @nullable */
  startDate: string | null
  /** @nullable */
  dueDate: string | null
  client_name: string
  project_name: string
  /** @nullable */
  start_date: string | null
  /** @nullable */
  due_date: string | null
  /** @nullable */
  createdAt: string | null
  /** @nullable */
  updatedAt: string | null
  /** @nullable */
  created_at: string | null
  /** @nullable */
  updated_at: string | null
}