// frontend/src/api/generated/models/storeProjectRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */
import type { ProjectPriority } from './projectPriority.ts'
import type { ProjectStatus } from './projectStatus.ts'

export interface StoreProjectRequest {
  /** @maxLength 255 */
  client_name: string
  /** @maxLength 255 */
  project_name: string
  /** @nullable */
  description?: string | null
  status: ProjectStatus
  priority: ProjectPriority
  /** @nullable */
  start_date?: string | null
  /** @nullable */
  due_date?: string | null
}